+++ haskell-src/Concordium/Cost.hs
@@ -121,13 +121,9 @@ removeBakerCost = 300
 configureDelegationCost :: Energy
 configureDelegationCost = 300
 
--- | Base cost for token holder transactions.
-tokenHolderBaseCost :: Energy
-tokenHolderBaseCost = 300
-
--- | Base cost for token governance transactions.
-tokenGovernanceBaseCost :: Energy
-tokenGovernanceBaseCost = 300
+-- | Base cost for token update transactions.
+tokenUpdateBaseCost :: Energy
+tokenUpdateBaseCost = 300
 
 -- | Cost for the execution of a token transfer.
 tokenTransferCost :: Energy